test: using the annotation
