+++ vaden/example/lib/src/dtos/date_time_parse.dart
@@ -1,0 +1,16 @@
+import 'package:vaden/vaden.dart';
+
+@Parse()
+class DateTimeParse extends ParamParse<DateTime?, String> {
+  const DateTimeParse();
+
+  @override
+  String toJson(DateTime? param) {
+    return param?.toIso8601String() ?? '';
+  }
+
+  @override
+  DateTime? fromJson(String? json) {
+    return DateTime.tryParse(json ?? '');
+  }
+}
