+++ vaden/example/lib/src/dtos/parse_dto.dart
@@ -1,3 +1,4 @@
+import 'package:example/src/dtos/date_time_parse.dart';
 import 'package:vaden/vaden.dart';
 
 @DTO()
@@ -15,18 +16,4 @@ class ParseDto {
     required this.createdAt,
     required this.updatedAt,
   });
-}
-
-class DateTimeParse extends ParamParse<DateTime?, String> {
-  const DateTimeParse();
-
-  @override
-  String toJson(DateTime? param) {
-    return param?.toIso8601String() ?? '';
-  }
-
-  @override
-  DateTime? fromJson(String? json) {
-    return DateTime.tryParse(json ?? '');
-  }
 }
